Explicit conversion to float added to reduce compiler warnings

+++ src/netcdf/create1_netcdf.c
@@ -76,9 +76,9 @@ Bool create1_netcdf(Netcdf *cdf,
       return TRUE;
     }
     for(i=0;i<array->nlon;i++)
-      lon[i]=array->lon_min+i*config->resolution.lon;
+      lon[i]=(float)(array->lon_min+i*config->resolution.lon);
     for(i=0;i<array->nlat;i++)
-      lat[i]=array->lat_min+i*config->resolution.lat;
+      lat[i]=(float)(array->lat_min+i*config->resolution.lat);
     if(cdf->state==ONEFILE)
       cdf->n=1;
     if(max(n,cdf->n)>1)
