+++ src/utils/cdf2clm.c
@@ -178,7 +178,7 @@ static Bool readclimate2(Climatefile *file,    /* climate data file */
           nc_close(file->ncid);
           return TRUE;
         }
-        data[cell*size+i]=file->slope*f[file->nlon*(i*file->nlat+offsets[index])+offsets[index+1]]+file->intercept;
+        data[cell*size+i]=(float)(file->slope*f[file->nlon*(i*file->nlat+offsets[index])+offsets[index+1]]+file->intercept);
       }
       else
       {
